noop refactor to quickfixes from refactors
Summary:
no-op refactor to make the next diff easier to review.

Abstract the logic of finding the refactoring for a given error into its own module. For context: {D53757680} introduced using refactorings as quickfixes by matching on a string name. But for errors with specific codes, like "DiscardedAwait", it's better to match on code )D53813494)

Potential future work: We could add a specific parser error for conditional await so we wouldn't need to match on strings for logic like that in {D53757680}. I don't think it's worth it at this point since we're considering removing the conditional await error entirely. And errors in general belong in later stages like elab rather than in the full-fidelity parser.

Differential Revision: D53813493

fbshipit-source-id: c5180e06634a34d0efdee0f593d7128a2a67639d