Fix race in WTCallback
Summary: As `interruptHandler()` does not steal the promise under the lock, if it is left with the last `self` reference the promise destructor races with the other callbacks that may be setting the value. To avoid this, we need to ensure that the promise is always moved from the callback, under the lock, when we intend to fulfill it.

Reviewed By: Gownta

Differential Revision: D53851683

fbshipit-source-id: b2235df25f0c1c1f3b0b34eba2e5392022a578cf