quickfix and refactoring for prepending 'await'
Summary:
Add both a quickfix and a refactor for awaiting an awaitable.

https://pxl.cl/4m2LV

The refactor is useful for when there is no error, but `awaiting` is something the user might want to do (and to know they can do).

The quickfix is for fixing an error quickly.

## For reviewers

Will we ever need to wrap the `await` expression in parentheses?
I didn't add such logic, but haven't been able to make the refactor/quickfix add the `await` in the wrong place. It seems `await` binds fairly tightly so that `await gen_int() + 1` parses nicely as `(await gen_int()) + 1`.

## Follow-on work

Ideally, the refactor/quickfix should make the containing function `async` if it isn't async already and convert the return type to `Awaitable<_>` as needed. I plan to do this in future diffs.

Reviewed By: enetsee

Differential Revision: D53813494

fbshipit-source-id: 13da2c09e60727fb96e4bc747f41441a2c26edb8