Be stricter about when to show quickfixes
Summary:
Here we're a bit more restrictive about when we show quickfixes:

https://pxl.cl/4kgXh

Before this change, we also showed quickfixes when the selection contained the error span:

https://pxl.cl/4kh51

With this change, we remove cases where we would show a quickfix but neither TypeScript nor Rust Analyzer would.

## Future Work

VSCode language experiences are not consistent in how they decide whether to show a quickfix: I saw different behavior in TypeScript and Rust Analyzer, for example. We're asking for guidance on this LSP issue: https://github.com/microsoft/language-server-protocol/issues/1897

Reviewed By: ljw1004

Differential Revision: D53568626

fbshipit-source-id: fe3ba85c19bd96fd491c60f362e8b0a17a1f2199