Properly account for changed request timeouts in Treadmill
Summary:
The Treadmill will kill the process (by sending SIGABRT to itself) if
it detects a request is "stuck". That is, if it's still running after
some multiple of Server.RequestTimeoutSeconds.

However, this neglects the fact that individual requests can change
their timeout. This means a request can set it's timeout very high,
run normally, then get killed by the Treadmill for being "stuck"
despite it running normally.

Fix this by taking into account both Server.RequestTimeoutSeconds and
the request's individual timeout when calculating the age limit.

While I'm at it, clean up some of the output for the Treadmill inside
crash-reporter.cpp. Namely, make it more clear on the crash dump that
the reason the server crashed was because the Treadmill killed it.

Add a new test case which reproduces the issue. It crashes before this
fix. Exercising this bug required the addition of a new intrinsic
which forces the Treadmill to check for stuck requests (normally it
only does so when a request begins, which is hard to reproduce in an
isolated test).

Reviewed By: jtwarren

Differential Revision: D53906513

fbshipit-source-id: a59d45b9a0000035f334279659b2555adc8f3eef