Fix shutdown race in HeapTimekeeper
Summary:
`HeapTimekeeper` is currently susceptible to the following race in `interruptHandler()`

- T1 cancels a future, fulfilling the promise
- T2 destroys the timekeeper
- T1 tries to schedule the cancellation of the timeout from the heap in the timekeeper

The race window is extremely narrow, and virtually impossible to hit when `HeapTimekeeper` is used as a singleton (all futures should be destroyed by the time of destruction), so handling this case was left as a TODO. This diff resolves the TODO.

To handle this, all the timekeeper state is moved to a `shared_ptr` that is shared with the futures. The shutdown protocol still drains the state on timekeeper destruction, so if any future is left holding the last reference it doesn't have almost anything to do.

Since all the state is encapsulated in a separate class, the vast majority of the implementation can now be PIMPL'd, as an additional minor benefit.

Reviewed By: Gownta

Differential Revision: D53857025

fbshipit-source-id: d8f5dee23c5091dd2f002fba024f8ebd6049511d