Break out class and injective constructor subtyping
Summary:
Context: we'd like to break apart the large set of mutually recursive functions into recursive modules to make the API clearer and enable easier modifications for error telemetry

Moves handling of class subtyping to its own module and exposes handling of subtyping with variance info to another module. We need this to be separate as subtyping for non-injective constructors calls into this.

Reviewed By: mheiber

Differential Revision: D53947827

fbshipit-source-id: b71a9bb7768c1593ab20c864ea9aeb1070d4a413