use fsame_slice in getFunctionPath()
Summary:
When you look up a toplevel function by name, it bottoms out in a
sqlite query against the function_paths table, whose entries are NOCASE.

Migrating the Facts DB is a pain, so we need a way to first log, then enforce,
correct case names, controlled by config. Once the migration is finished
we can change the DB itself.

To do this, AutoloadDB::getFunctionPath() now returns a list of (path, function)
pairs containing the actual function names from the DB. Filter this with a new
comparison wrapper function fsame_slice() that either logs or compares exactly
based on the Eval.LogFsameCollisions config setting.

The StringPtr::fsame_slice() wrapper enables a non-hhvm implementation
for the test suite, which is not allowed to depend on runtime_core.

Reviewed By: ricklavoie

Differential Revision: D53779050

fbshipit-source-id: e808b6415efb78bc8cf610c841bc26029c560e57