increase timeout in the FileDecls_from_path_repeated.php test
Summary:
This has been causing sporadic failures in automated test runs (often when hhvm is compiled with slower build modes such as dbgo builds).
In theory we shouldn't rely on timing during tests (tests can run on a machine that also runs other cpu-intensive things and thus are not guaranteed to finish in a few seconds). But intead of deleting the test, we'd just let it pass to unblock the automated release pipeline.

Reviewed By: ricklavoie

Differential Revision: D53955020

fbshipit-source-id: e65e9b92bf8d242b9026095149256605d9b23e48