Only do this trick for non-primitive types
Summary:
The previous diff (D53824266) broke some builds. For example:
```
buck2 build //datainfra/dmeta/common-security:dmeta-common-security_lib-java11
```

The issue is that some people are using constants in `switch`-`case` statements
(of course this is kosher). The problem with that is that there is actually a
distinction between `static final` variables, and "constant expressions"
(defined [here](https://docs.oracle.com/javase/specs/jls/se8/html/jls-15.html#jls-15.28), and only the latter can be used in a `case` statement.

Since only primitives and strings can be used in those locations, we disable
this transformation for such types. The benefits of this transformation still
work, since these types are exactly those that require no fancy methods to
initialize :)

Reviewed By: bmdahmen

Differential Revision: D54043616

fbshipit-source-id: 8e4d39e9db399322ba26058f4d745f8893fd4773