Move handling of transitivity and top-level search to separate modules
Summary:
Context: we'd like to break apart the large set of mutually recursive functions into recursive modules to make the API clearer and enable easier modifications for error telemetry

Moves handling of transitive props to a seperated module. Also moves the top-level entry point for 'telling' subtype constraints to it's own module - this simply chains a call to `Subtype.[...]` with a call to the recursive `Subtype_trans.prop_to_env`.

Differential Revision: D54000177

fbshipit-source-id: 1e1e253f0553469db70942bba784a60057c1fd6d