add module information to gconsts
Summary:
When checking whether a module newtype is visible we check if the type being used is defined in the same module
https://www.internalfb.com/code/fbsource/[8331eb2974fac88d05acde72b682d38de8e7aa5b]/fbcode/hphp/hack/src/typing/env/typing_env.ml?lines=718-723
Today global constants don't carry module information with them and hence don't know which module they belong to. This diff augments gconsts with a module field so module newtypes could be expanded correctly.

Reviewed By: vassilmladenov

Differential Revision: D53976788

fbshipit-source-id: da2fb1651918923ba26e7885e08a0ba50e88dde0