Frozen: Allow heterogeneous comparisons for maps
Summary:
PROBLEM
Given a frozen map, the only way to do a lookup is using a frozen key. A relatively common use-case is to have a frozen map, but a key retrieved from the user (or somewhere else) - Which implies that the key itself isn't frozen.
In case the key isn't directly convertible to a View (such as a user defined structure), the key needs to be explicitly frozen, before it can be used for lookups. This adds a penalty that is avoidable, since only the hash, equality and comparison operators need to be defined for the key.

SOLUTION
Similar to how C++20 introduces `is_transparent` on comparators, we allow heterogenous lookups on Frozen maps. The TL;DR is given a `K`, the map would return an element whose key compares equivalent to the given `K` argument.
Note that there are types that are convertible directly to a `Layout<T>::View`, since C++ would prefer matching to the `K` override over the conversion, we explicitly disable that function if there is an implicit conversion between `K` and `Layout<T>::View` (in order to not break existing logic).

Read more: https://fb.workplace.com/groups/thrift.frozen/permalink/7216711045086997/

Reviewed By: vitaut

Differential Revision: D53889850

fbshipit-source-id: 96e25d84c78be7bdae53e429147b0f8bf8e893e5