intern TypedValue::LazyClass
Summary:
Migrate hhbc::TypedValue::LazyClass(Slice<'arena, u8>) to StringId, which is
an interned utf8 string, since hack source text and thus symbols are unicode.

The main change is in hackc/hhbc/typed_value.rs where LazyClass takes a StringId.
This kills the lifetime parameter on TypedValue, Adata, AdataState, and some funcs.
The rest is adjusting to the new reality.

Reviewed By: aorenste

Differential Revision: D53417045

fbshipit-source-id: 5ab39342fa82d44c5ac1c64976b3d02e4842089e