Avoid generating expression-dependent type when it's safe to do so
Summary:
Expression-dependent types are necessary only for classes that EITHER define abstract type constants (which can be realized by distinct types in subclasses) OR have methods that make use of `this` in their signatures (because these can be realized by distinct types in subclasses). Unfortunately, there is nothing at the class level that flags this property, but for well-known built-in classes such as `vec` and `dict` we can avoid generating an expression-dependent type.

Comments welcome on how we might generalize this.

Reviewed By: mheiber, madgen

Differential Revision: D53920324

fbshipit-source-id: e68445d1155cbc40c364faba7a0536da25e4a6f0