Revert D54017026: Multisect successfully blamed "D54017026: Combine Ast and Schema structs" for one test failure
Summary:
This diff reverts D54017026
D54017026: Combine Ast and Schema structs by iahs causes the following test failure:

Tests affected:
- [cogwheel:cogwheel_windtunnel_e2e_run_slcm#main](https://www.internalfb.com/intern/test/562950069774915/)

Here's the Multisect link:
https://www.internalfb.com/multisect/4389250
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Differential Revision: D54112811

fbshipit-source-id: 6e391902aea15bad8fe323d593eead8158b58aa9