Add a comment to TransCFG logic
Summary:
This was confusing to me when I first read it. Adding a comment.

Before we create a profile translation, we create a `SrcRec` for it that we of course store in the `SrcDB`. When we are in a
`jit-serialize` mode, the `SrcDB` on the machine we are deserializing on will not have the profiling translations we are looking for since we collected them before we restarted.

Let me know if this isn't quite correct.
Is there any state that persists between restarts?

I had a thought about refactoring this a little so we just do this computation in one place whether we plan to serialize the
profile data or not. We effectively would just need to do the calculation after we were done profiling.
But this would require adding something to RTA https://fburl.com/code/kv89u5ps for serialize/deserialize case, and something to RetranslateOpt for the case where we are doing func-at-a-time retranslation. This feels like a worse special-case than what
we do now, so I left it like this.

Reviewed By: ricklavoie

Differential Revision: D54108377

fbshipit-source-id: b0860ab095e792f3e74d41a274d99985411bf75b