Fix per file telemetry
Summary:
For per file typechecking telemetry, we run the typechecker twice once to get the decls warmed up and a second time to make a measurement that tells us about which file is compute expensive.

This was broken as a result of adding a new named parameter to the `process_file` function. The second invocation became a closure that never got computed.

This diff adds the missing argument to fix the error.

Differential Revision: D54119289

fbshipit-source-id: 38dc45fa00e203a34eaea06e51c20b1365da5968