fix incorrect use of if-constexpr in operator_delete
Summary:
There was a use of `if constexpr` in a non-template function to dispatch arguments, based on whether the forwardee would be invocable with those arguments. But this was not in a template, so `if constexpr` is not a suitable technique for that site.

> Outside a template, a discarded statement is fully checked.
> https://en.cppreference.com/w/cpp/language/if

Fix by moving the `if constexpr` into a template.

Reviewed By: interwq

Differential Revision: D54121955

fbshipit-source-id: 7b8f2e3ac0f8da71e1c33ef03489ec919f5fa957