add method to get iobuf for writing
Summary: This diff adds a new method to the MessageWrapper class in the Thrift library. The method takes a function pointer that returns a string view of the message buffer and a length parameter, and returns an IOBuf that wraps the buffer as a value and has the specified length.

Reviewed By: fadimounir

Differential Revision: D54083501

fbshipit-source-id: c2af993149936e5f570151cff76bbdf1440be107