let Singleton skip scheduling atexit on iOS
Summary:
Global variables initialize themselves with a call to `atexit` to schedule destructors at shutdown time. `Singleton` simulates this behavior closely.

iOS sometimes injects a call to `exit` into the main event loop. This is wildly incorrect from the perspective of POSIX and the entire history of how to write software. But iOS nevertheless does it. The call to `exit` calls all of these scheduled destructors. To be correct, either all `atexit` handlers must run in the order scheduled or none must run.

We have many instances of global thread-pools or other global objects with destructors that can wait. When iOS injects a call to `exit`, there is nothing that previously would have signaled anything else to stop, and so all these global objects continue on without stopping and their destructors wait indefinitely. This is a problem in practice.

In practice, most cases of of this are done with `Singleton`. This creates a workaround for the problem by skipping the call to `atexit` in `Singleton`. The same problem could occur for global objects done as normal global variables, but with no workaround.

Reviewed By: yfeldblum, ot, fabiomassimo

Differential Revision: D46993058

fbshipit-source-id: 1c7b8b7fc469a3c374331fbd3f55669068a88ab5