complete hint pretty-printer
Summary:
The `displayType` field in `hack.TypeInfo` corresponds to the pretty-printed hack hint, and is eventually rendered by codehub, together with type xrefs.

The pretty printer used to compute  this is incomplete. It falls back on a different  pretty-printer (`get_type_from_hint_strip_ns`) for the unhandled cases.

As a result,
1. Some hints aren't printed properly (see linked task)
2. Some xrefs are missing
3. Some large types are truncated (get_type_from_hint_strip_ns rely on a hack pretty printer which truncates long types)

This diff gets rid of the fallback and implements all the missing cases.

## Discussion

`get_type_from_hint_strip_ns` isn't what we want to index hints. It is based on hack pretty-print in `Typing_print`.
1. it' s not able to compute the  xrefs
2. it converts the *hint* (i.e. the user types) to the "internal" hack *type* and pretty print those internal types which are richer. As a result, the string representations contain extra stuff which weren't visible in the hint, and that we don't want to display in codehub.
3. we can't easily modify it to our liking because it's used in other parts of the code base.

There's actually a hint pretty-printer that could have been used [`Hint_print.pp_type_hint`](https://fburl.com/code/hq9j1e2x). But it's not ideal either for similar reasons (difficult to instrument to add xrefs without rewriting it + not the format we want, e.g. it adds newlines, doesn't get rid of the supportdyn annotations).

As a follow-up, I'd like to get rid of other uses of `get_type_from_hint_strip_ns`.

Reviewed By: nhawkes

Differential Revision: D54000838

fbshipit-source-id: 71ed82b2ae275d81d9d675d472720af4d0751b9b