More graceful handling and actionable error of remote_old_decls_ffi not found
Summary:
This error is expected to only happens during internal Hack team development when a member of Hack team has build with `-c hack.distc=0`.

No need to crash: fail gracefully and make the problem as discoverable as we can with `invariant_violation_bug`, which will be loud in logs.

Reviewed By: nt591

Differential Revision: D53948341

fbshipit-source-id: 8d5d610ff52d660f4e1ce4b296c9f8f8f032833d