Suppress information about existentials where it isn't useful
Summary:
`hh` encodes existentials using `Tdependent` and `Tgeneric`. When rendered, this  has the slightly confusing form of

`<expr#\d+> as ...`

In most cases when this type of this form appears in an unsatisfied subtype constraint it occurs on only one side. In these cases, the existential itself cannot be directly involved in the error.

This diff suppresses the rendering of the existential in these cases and adds a flag the `User_error.t` which lets us track where we have made use of the suppression

Reviewed By: mheiber, andrewjkennedy

Differential Revision: D54198160

fbshipit-source-id: ea3c5574cfc6b3a3769a33fc510644bfa5fca0fe