go codegen cache for munge_ident
Summary:
Add a codegen cache for munge_ident.

Use it initially in function codegen for getting a unique param name, but it may also be used in more places. Also switch function codegen get-unique-name to do a linear search on a vector rather than a set-contains search, since it is expected that in each separate case there will be few enough items.

Reviewed By: awalterschulze

Differential Revision: D54198643

fbshipit-source-id: b83e765c5fb2816f3fc1660356d7c2c95282f306