copy in generated rpcmetdata
Summary:
To avoid a circular dependency, we temporarily copy the generated code for rpcmetadata into the thrift library
rpcmetadata is Generated code, which imports thrift library
rocket implementation is in the thrift library, but needs to import rpcmetadata

We are working on a solution in parallel, but we don't want to block rocket implementation forward first.

Problems:
1. The thrift library is currently in one folder, because buck paths and go paths do not match. "thrift/lib/go/thrift" != "github.com/facebook/fbthrift/thrift/lib/go/thrift", which means that open source users cannot use a new folder as the import would differ from the internal version. This is why the imports in the generator are configurable and different for users inside and outside of facebook.
2. We need to figure out a way to set namespaces with `.` characters, as in `github.com/...`
3. The circular dependency mentioned above
4. Where to check in generated code for rpcmetadata and any, so that it can be used by both buck and go build.

This pragmatic solution, gets around all these problems.
We still think this would be good to solve in future, but for now adding rocket support is the priority.

Reviewed By: podtserkovskiy

Differential Revision: D54198892

fbshipit-source-id: 53a540db932b6e1dc4e5b322045d0764740628ca