Start splitting up `simplify_subtype_i`
Summary:
`simplify_subtype_i` is a beast of a function (over 2300 loc) largely because of nested helpers. This diff starts to break it up by moving out the code dealing with `constraint_type` in supertype position.

(Long term, it would be good to refactor constraints, propositions and the inference env to avoid dealing with non-subtype constraints in subtyping code. I chatted with CatherineGasnier about this a while ago and it was done to make transitivity easier but I don't think it is necessary)

Reviewed By: madgen

Differential Revision: D54070339

fbshipit-source-id: fd6176462a1b588de789bff8a206f560f32d44bb