Eliminate duplicate cmd names
Summary:
Adds logic in `build_fixtures.py` (and `fixture_utils.py`) to enforce that the names added to individual `cmd` lines in the previous (stacked) diff are unique for a given fixture.

Subsequently updated the newly introduced names to ensure uniqueness.

Note that, other than enforcing the uniqueness property above, these names are not yet used for any other purpose. The next diff adds logic to generate the output of each cmd line in a dedicated folder based on the (now, unique) name.

Differential Revision: D54235351

fbshipit-source-id: a757ee969aa0cb6148a0b5b29b0c72ea3f66aeab