fix well-formedness checks for internal methods in module level traits
Summary:
If an internal method is defined in a trait, then the trait must be either internal or annotated with `<<__ModuleLevelTrait>>`.  This is enforced by HHVM at runtime during class loading.

In repo mode the existing check was not handling correctly the case where an internal method is defined in a module level trait that it used by a non module level trait.  Consider

```
<?hh
module A;

<<__ModuleLevelTrait>>
public trait T {
  internal function bar() { echo "bar in T\n"; }
}

public trait T1 {
  use T;
}

class C {
  use T1;
}
```

(see `internal_methods_01.php` for a complete example).  After HHBBC inlining, trait `T1` was considered to define the internal method `bar`, and HHVM raised an error because it ignored that `bar` was originally defined in a module level traits (and as such its runtime representation is extended with the information needed to correctly perform runtime module boundary checks).

HHVM does not track the original trait where a method, before HHBBC trait inlining, was originally defined.  However the fix in this diff relies on the query "has this method been inlined from another trait" (`method->isFromTrait()`), and exploits the following logic:

- in sandbox mode, traits are flattened at class load time.  The check "is this an internal method defined on a public, non-module level trait" is performed on each method only once, when checking the methods defined in a trait.  It is not performed on methods copied across `use` definitions.  The condition `method->isFromTrait()` is always false, and the overall logic is correct;

- in repo mode we must distinguish two cases
  - an internal method `m` is defined in an internal or in a module level trait `T`, used by other traits.  In this case HHBBC inlines the method in the traits that use `T`.  At class loading, HHVM will perform the well-formedness check on the method when loading trait `T`.  When loading a trait `T'` that use `T`, the check `method->isFromTrait()` for method `m` will return true, and the check won't be repeated.
  - an internal method `m` is defined in a public, non-module level trait `T`.  In this case HHBBC will detect the issue and will not inline `m` in the traits that use `T`.  HHVM will report an error when loading class `T`.

Differential Revision: D54118139

fbshipit-source-id: 306bf6d622840884e54906b528fc933e86569257