add static prop data in objprof
Summary:
This is second attempt to achieve this. The first attempt D53795396 required us to add that new sprop mappings into s_handletable inside bindOnLinkImpl. The problem was that the approach may cause minor regressions AND was not thread safe.

This approach simply traverses through classes that have sProps initialized and accounts for those

Reviewed By: ricklavoie

Differential Revision: D54188862

fbshipit-source-id: 8f8c3e1bf1f15a440a4ba64202366722e4022df2