introduce recordSboxEventWithTimingAndError to unify the idea of perf and error logging
Summary: Instead of having one row for operation perf and one row for errors, we want to move to a world where every operation gets a single row and error information is either present or not depending on if it succeeded. This stack takes a few steps in that direction, moving all the calls in FactsLogger to record timing and error information instead of just timing info. We also get rid of one manual error logging callsite which is duplicative.  We will convert the other manual error logging callsites to general purpose perf/error logs over time.

Reviewed By: edwinsmith

Differential Revision: D54128135

fbshipit-source-id: 2f0de44c2d467b3a594846ad5cad58f409a1af5d