Add local iterator optimization to hackc
Summary:
Attempt to generate LIter{Init,Next,Free} bytecodes in HackC when iterating local variables. When we see that a `foreach` loop is iterating over a local, and that the local is either not mutated during the loop, or that any mutation to the local is only preformed at the current iteration key (and is not an unset), then we can use these forms of the iter bytecodes which will avoid holding an extra reference.

This is particularly important for the HSL where certain functions will attempt to mutate an array in place while iterating it. This is designed to avoid a copy but on sandboxes without this optimization we still incur the full copy.

Reviewed By: edwinsmith

Differential Revision: D53687642

fbshipit-source-id: 49a981e3af740bdb4bcdf379da4c780c7c25af15