Introduce CGetL->PushL optimization to hackc
Summary:
This diff adds a new analysis pass to hackc that does both forward (to determine which locals are known to be initialized) and reverse (to determine which locals are anticipated to be live) dataflow analysis. It uses this information to perform the following transformations:
 1. Rewriting `CGetL` instructions to `PushL` when the local is known to be initialized and the `CGetL` is the final read of the current value of the local.
 2. Re writing `BaseL` to `PushL; BaseC`, in addition to the rules from (1) the member operation must also be a `QueryM`
 3. Upon seeing any instruction which reads from a local if that instruction is the last such read from a local an `UnsetL` is inserted following the instruction.

The (1) and (2) optimizations are designed to mirror the behavior of hhbbc, while (3) is designed to be a lighter weight version of hhbbc's behavior (unsetting unused locals possibly containing CoW values) as type information is not inferred.

Two different levels of aggressiveness were added for these optimizations:
 1. Hack.Lang.OptimizeLocalLifetimes - enables/disables the optimization
 2. Hack.Lang.OptimizeParamLifetimes - causes parameter locals to not be optimized ensuring they remain available for `debug_backtrace`

Reviewed By: edwinsmith

Differential Revision: D52965874

fbshipit-source-id: f30fea3a6a6b3e12842852d9a2fd101f13882d2a