run fb303 server before warmup (v2)
Summary:
I caused S390663 with this before because it turns out the fb303 server will try to use the HttpServer that has not yet been initialized at the beginning.
I didn't find this issue before because it's unlikely to trigger this condition. Something must take a lot longer in prod than on my devserver which made it more likely.

To account for that, I've been reproing/testing this change with a sleep(10) added (see test plan).

Reviewed By: 3ach

Differential Revision: D53241042

fbshipit-source-id: 9ba8c135e053539c4e2723c6903b95366bbbc287