Re-add Type canonicalization in make_local due to case types
Summary:
In D50826876 I removed the logic to "resolve" Types from extern-worker
jobs because it was no longer necessary. The Types produced from the
InitTypesJob were guaranteed to be in canonical form. Unfortunately,
the introduction of case types broke this. Computing the initial
return type for a function which returns a case type involves
intersections. However, each shard in the InitTypesJob does not have
complete class information. Therefore the type computed in each shard
may not be canonical. Since all of the Type logic assumes
canonicalization, this causes subtle breakages.

Re-add logic similar to what was previous removed. Call it
"unserialize_type" because of how it will be used in a later diff.

Reviewed By: vassilmladenov

Differential Revision: D54349543

fbshipit-source-id: d27fd64ad6c8b445261eda50fa952017c527bc9f