Fix unsoundness introduced by D44294183
Summary:
As title says, an earlier diff D44294183 unfortunately introduced an unsoundness in the way that we type check method calls through like-type receivers, and function calls through like-type function types. For example:
```
class C {
  public function foo():int { return 3; }
}

function testit(dynamic $d):void {
  // $x has type ~C
  $x = HH\FIXME\UNSAFE_CAST<mixed,C>($d);
  // We expect $y to have type ~int
  $y = $x->foo();
}
```
The type of `$y` above is inferred incorrectly to be `int`, not `~int`.

The fix is to wrap the result with a like type whenever `dynamic_func` is `Like_function`.

Reviewed By: dlreeves

Differential Revision: D53275932

fbshipit-source-id: 9c940801a3eb4cf1a3c4288349a8259e7585029e