Properly Instantiate this_ty for Enum Class Members + Gating
Summary:
# Context

When the enum class with type constants feature was added, this allowed the type `this` to appear as part of the type for an enum class member. We were not properly instantiating the `this` class, since we were not providing an instantiation when localizing the type from an enum class member. This would default to the `this` type of the enclosing class which is incorrect.

# Mitigation

The fix is to provide an instantiation for the `this` type when localizing. In this case we use the type of the class passed into `EnumClassLabelOps.expand`

Reviewed By: viratyosin

Differential Revision: D54439754

fbshipit-source-id: 67634ae0bcb70ba2c1889fb424fa1c4c5175b922