Fix coro::Task co_yield destroying locals on a different executor
Summary:
If a normal result is returned (co_return) or exception is thrown (throw) all the locals (e.g. guard) are destroyed before final_suspend coroutine is awaited.
If co_yield is called, all the locals are not destroyed before final_suspend coroutine is awaited (since final_suspend is called inline from co_yield), and they will only be destroyed when the coroutine frame gets destroyed.
The coroutine frame destruction happens from within an Awaiter (which is owned by whoever started the coroutine), when the result is extracted from that Awaiter (which may happen on a different thread).

Given the widespread usage of co_yield co_error/co_result, we have to make this pattern safe. This diff injects an additional coroutine whenever TaskWithExecutor is awaited from another Task to make sure that all the locals from the TaskWithExecutors are guaranteed to be destroyed on the correct executor. This will be a performance regression, but correctness is more important here.

Reviewed By: yfeldblum

Differential Revision: D54130736

fbshipit-source-id: b1a737dcac3f7e2370fb6f0bebac55aaa57cb8d6