fixing optional arguments in traits
Summary:
Hack function with optional N arguments generate N+1 Textual functions (all cases were 0, ..., N values are given). But currently this generation is performed after the addition of the special `self` parameter in traits methods. This is wrong since this argument will only appear in the last Textual function while we need it in all N+1 versions.

We fix the problem by following what is done with ReifiedGeneric args, i.e. doing the parameter addition after the generation of the N+1 functions.

Reviewed By: aorenste

Differential Revision: D54491165

fbshipit-source-id: 9fce863227fe57b9e24fc521a3805001a505da28