intern SA
Summary:
Use BytesId for opcode SA immediates (string literals).

Future: We might want to split SA into two types (SA, SY?) with one of them being for symbols
we know must be utf8, and SA for things that must be hack string literals (non-utf8).

This is the last opcode immediate that needed `'arena` so the cleanup extends to Instruct
and InstrSeq.

Reviewed By: aorenste

Differential Revision: D53853722

fbshipit-source-id: 29360dc66383a281271ffd6486785e43074032f6