Mitigation fix for leaking internal type constant projections
Summary:
We allow inheritance of members from internal classes and traits into public ones, even if their signatures include internal types, but when localizing such internal types we pretend that a `Tclass` is a `Tnewtype` i.e. is opaque. Unfortunately this breaks with type constant access, so (for example) `self::TC` in an internal class, inherited into a public class, results in a `Tany` i.e. a soundness hole. Here is an example:
```
//// module_A.php
<?hh
new module A {}
//
//// module_B.php
<?hh
new module B {}
//
//// A.php
<?hh
module A;
internal class Base {
  const type TC = string;
  public static function meth1(Base::TC $_):void { }
}

final class Derived extends Base {
}
//// B.php
<?hh
module B;
function main(): void {
  // Should be rejected!
  Derived::meth1(1);
```
The preferred solution is to ban public classes inheriting signatures that contain internal types in their signatures, and require the user to ensure (e.g. via module newtype) that the signatures are denotable in the inheriting class.

But to mitigate the problem for now, we can ensure that a type projection off such a "fabricated" `newtype` is itself rendered as another fabricated `newtype`. It's important that the name of the type is unique to the type constant projection, which we ensure by mangling.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D54533697

fbshipit-source-id: ccc56c829ed7e2a50b097bb73bbb459a8632376b