Fix potential deadlock in MessageQueue
Summary:
When using relaxed or batched notifications the queue may become full before we have the chance to send the notification. If nothing else unblocks the loop (triggering the before-loop callback that drains the queue) the producer may deadlock.

In practice we likely never observe this because there is always something eventually unblocking the loop (for example, scheduled timeouts), but we should not rely on that.

Reviewed By: lenar-f

Differential Revision: D53534737

fbshipit-source-id: 497e17282df06add4d5c57305fa8ad905651f297