Use AllocationColocator in ContextStack
Summary:
For performance reasons, `ContextStack` coalesces multiple objects into one allocation. This diff replaces the manual colocated allocation (and related ugliness) with `AllocationColocator`, which encapsulates this under an abstraction.

Here's the allocation scheme, in ascending address order:
* `create`
  * `ContextStack`
  * `N`×`void*` context pointers
* `createWithClientContext`
  * `ContextStack`
  * `EmbeddedClientRequestContext`
  * `N`×`void*` context pointers
* `createWithClientContextCopyNames`
  * `ContextStack`
  * `EmbeddedClientRequestContext`
  * `N`×`void*` context pointers
  * `serviceName` string
  * `methodName` string

This diff preserves this scheme. However, there are some small changes:
* `ContextStack` object is larger because it stores a `void**` which is the start of the context pointers array
* `ContextStack` object is larger because it stores an `AllocationColocator<>::Ptr<EmbeddedClientRequestContext>`.
* The alignment of the `malloc` is different. It used to be `alignof(ContextStack)` but will now be `__STDCPP_DEFAULT_NEW_ALIGNMENT__` (16 bytes in practice). Very much not a concern.

None of these are of concern — the performance bottleneck is not memory, it's the extra CPU from `malloc`s. There is still exactly one `malloc` per `ContextStack` after this diff.

Another fun bug...

https://www.internalfb.com/code/fbsource/[07ab1d108e3184165e168743f2bd205c8658bba9]/fbcode/thrift/lib/cpp/ContextStack.h?lines=35-44

`ContextStack::create*` return a `unique_ptr` with a custom deleter. That means that the specialization of `std::default_delete` is unused (LOL).
https://www.internalfb.com/code/fbsource/[07ab1d108e3184165e168743f2bd205c8658bba9]/fbcode/thrift/lib/cpp/ContextStack.cpp?lines=270-293

It just so happens that the colocated objects (`EmbeddedClientRequestContext`, `void*` pointers, C-strings) are all trivially destructible — we just got lucky here. So this diff also removes all that.

Reviewed By: sethdelliott

Differential Revision: D54439166

fbshipit-source-id: 64c535d503a04dbbe2425733ea7fe657c13e22b3