Don't pass redundant `internal_ty`s
Summary:
When breaking up some of our functions, it became clear that we were using unexpanded internal / local types in recursive  calls.  This is almost certainly unintentional.

To have the minimal possible change, the previous steps in the refactor retained this behaviour. Here, we eliminate it and directly reconstruct the `internal_type` if it is needed in a recursive call. This ensures we always propagate the expanded type in recursive calls

Differential Revision: D54580728

fbshipit-source-id: 2e2a191f5da31a8fcceae130c344ad7ef1a8e783