Fix generics-related unsoundness with fake opaque version of internal types
Summary:
Distinct instantiations of `internal` generic classes are treated as compatible outside the module where they are defined. This is unsound. See the example test.

Fix is to include type arguments on the `Tnewtype` that's produced during localization. But we need to disable the generation new types during the "multiple instantiation" check, as this breaks through module boundaries in order to do localization and subtype checks.

Differential Revision: D54465555

fbshipit-source-id: 9449436385b0750ac694d0b376310750fdbbed76