Add SubselectionServerListFilter class
Summary:
This diff adds a new class SubselectionServerListFilter which will act as a filter on top of the smc fetch result.

We make use of proxygen's existing RendezvousHash implementation, but add a new build function to support pre-calculated node hashes, since folly::SocketAddress provides its own `hash()` function.

Then when we do subselection treat the revproxy's own id (tw task handle) as the selection key. I've decided to include the task version as well, so that a task restart will result in a new subselection.

Reviewed By: kwaugh

Differential Revision: D53593585

fbshipit-source-id: adc2d3cdcdc82a3d75892c77270b59c509ff7960