Check whether generic in is_typedef
Summary:
For `T::class` where `T` is a reified generic, we were relying on naming provider to not recognise `T`. This is not buggy as a generic type name `T` is resolved different to a defined type name `\T`, but it is still brittle.

This diff first checks if we have a generic at hand when we query if something is a type def.

Reviewed By: vassilmladenov

Differential Revision: D54602638

fbshipit-source-id: e6a8af422db98ab0c95793cb429c04488301956b