Add TProcessorBase::IgnoreGlobalEventHandlers
Summary:
Consider `TProcessorBase`'s constructor (base class for `AsyncProcessor`):
https://www.internalfb.com/code/fbsource/[ea57c3a57038]/fbcode/thrift/lib/cpp/EventHandlerBase.cpp?lines=55-61

We end up reading the set of globally registered event handlers (via `addProcessorEventHandler`). This is a problem for "wrapper" `AsyncProcessor` implements such as `MultiplexAsyncProcessor` and `PreprocessingAsyncProcessorWrapper`. These implementations also pull in the globally registered set but because they delegate `executeRequest`, they end up unused. This diff just allows subclasses to avoid the extra work.

Another effect is that calling `addEventHandler` on the "wrapper" ends up adding to the wrong list of handlers (the wrapper instead of the delegation list). That will be fixed in a later diff.

Reviewed By: sethdelliott

Differential Revision: D54598767

fbshipit-source-id: b482f958dab532ceda0b5817ae1eed8fd8362b9f