remove dependency on folly async_transport from fizz protocol_params
Summary:
Continued work to remove openssl dependency from core fizz code.

Currently the fizz state machine has a dependency on folly:AsyncTransport, which in turn depends on OpenSSL. The dependency on AsyncTransport is through the `folly::AsyncTransportWrapper::WriteCallback*` in `fizz::AppWrite` (in `protocol/Params.h`).

This WriteCallback is a type specific to the IO layer and the state machine layer should not depend on this. So this diff replaces the callback field of AppWrite with a `void* token`. The IO layer will take care of handling the token - it knows to cast it to a `WriteCallback*` - and the SM layer can remain agnostic, passing the token to an IO defined callback as necessary.

Reviewed By: mingtaoy

Differential Revision: D52983116

fbshipit-source-id: 751ea559fc4df1fcd1707a6094f6de4498870eaa