use preconfigured threadpool to improve perf vs
Summary:
The old model was pretty garbage. I was starting a threadpool on each
web request. Now, I lazily create a threadpool as part of the extension and
pass a KeepAlive token to any function that needs to run work on the pool.

I was getting segfaults without adding the `moduleShutdown` code since the
threadpool was just hanging around in the background and work would end after
the process shut down.

Differential Revision: D54568202

fbshipit-source-id: 668f42e82bd4c6f1d79eaf5cd4110a2f3b131c6f