Correct typing for comparison operators.
Summary:
The implementation of these operations checks for the type of the object, and return the appropriate False/NotImplemented value, as described in https://docs.python.org/3/reference/datamodel.html#object.__lt__

But since the typing imposed the same typing between `self` and `other`, pyre is complaining when comparing a structure to an `Optional` of itself.

See D54495137 for an example.

Reviewed By: ahilger

Differential Revision: D54639095

fbshipit-source-id: 337df647440fb16cb83a9ec52661bee225f3e119