Turn off meth caller optimization in tests
Summary:
This cleans out the tests that reference EmitMethCallerFuncPointers; this option is off by default and needs some work to avoid making breaking changes (see D54522173 tests). The next diff prunes it.

Since the option is old, the tests have diverged somewhat, e.g. the old test expectation is `object must be an instance of (D)`, but without the optimization we add `instead it is (B)`

Reviewed By: ricklavoie

Differential Revision: D54696623

fbshipit-source-id: c750b9f797f2c06728f7fcc81ee9488b4d38f5c8