Uniform interface for top-level constraints
Summary:
To break the recursion from non-subtype constraints back into `Subtype` via `default_subtype`, we want to inline handling of common cases like union-l, intersection-l, generic-l. However, there are some heuristics applied in those cases and we'd like to avoid copy-pasting.

To do so, we will lift those cases into a helper module and supply the function with which to recurse.

To make this easier, we first provide a uniform interface for each of our top-level constraints in terms of a (shared) left-hand-side and a specific right-hand-side for each constraint

Differential Revision: D54730199

fbshipit-source-id: 34b4cf69ad130927dce058f0a9b31cd84ca33075