Refactor Code_action_types so we don't have two modules named 'Quickfix'
Summary:
We had Quickfixes.ml and Code_action_types.Quickfix, which was empirically confusing (a teammate was confused).

Refactor so there is just `Code_action_types.t` with a tag to distinguish whether we're working with a Quickfix or a Refactor.

This change also simplifies the code for producing code actions, since Lsp.CodeAction (which corresponds closely to the LSP spec) uses a `kind` field to distinguish between quickfixes and refactors

## Review Advice

The main changes are in code_action_types.ml and server_code_actions_services.ml. The rest just follows.

## Pre-land Checklist

- land this *after* hverr's changes if it looks like the code would otherwise cause merge conflicts for Henri's work in progress

Differential Revision: D54800844

fbshipit-source-id: 6667b4328213c8a899f68a9c414c0e32c194df0a