Lift common functions from `default_subtype`
Summary:
To break the recursion from non-subtype constraints back into `Subtype` via `default_subtype`, we want to inline handling of common cases like union-l, intersection-l, generic-l. However, there are some heuristics applied in those cases and we'd like to avoid copy-pasting.

To do so, this diff lifts those cases into a helper module and supply the function with which to recurse.

Differential Revision: D54771343

fbshipit-source-id: ac2bca7b2b14ca6387db4bfe357cd505969f5a0d