Add support for incrementally-enabled dedicated output directories
Summary:
In a previous diff (D54322117), we enabled unconditional dedicated output support (to prevent fixture commands from clobbering each other's output). While functional, the resulting change is massive (7000+ files) and difficult to safely review (let alone get through CI and commit).

This change enables the same support, but conditions it on the existence of a file (`ENABLE_DEDICATED_OUTPUT_DIR`) in the fixture directory. This enables incremental (and parallel) rollout of the feature (with smaller, easier to review changes), in a way that keeps both `build_fixtures.py` and `fixture_test` in sync.

Differential Revision: D54811021

fbshipit-source-id: a131730adcd975a368157f9bec5a8008f857ce3a