Fix int64 handling sqlite3
The extension was using the sqlite3_*_int API, which
would truncate 64-bits int to 32-bits on some code
paths. The fix is to use the 64-bits API variants
constantly throughout the extension.

Added a test that manipulates int64 that are not
representable as int32 to make sure we handle
those correctly.