Introduce TypeSwitch Constraint Type
Summary:
# Context

Hack's current refinement system depends on generating negated types and intersecting them with the type of the variable. I.e.
```
Γ[$x -> Tx & T]  |- e1
Γ[$x -> Tx & !T] |- e2
----------------------------
Γ |- if($x is T) e1 else e2
```

This typing rule leads to incompleteness under two conditions, when `Tx` is a type variable and when `T`'s negation is approximated.

To resolve this I propose introducing a new typing rule:
```
Γ |- $x : TypeSwitch(P, T1, T2)
Γ, $x: T1 |- e1
Γ, $x: T2 |- e2
----------------------------
Γ |- if($x is P) e1 else e2
```

In this rule we introduce a new constraint type `TypeSwitch` which is used to encode flow sensitive refinement. Here `P` denotes a type predicate, which represents a runtime type test. `T1` represents the set of values that will pass the predicate at runtime, while `T2` the set of the value that will fail the predicate. Assuming we can define subtyping rules for this constraint type, then we can fill in `T1` and `T2` with type variables that the constraint solver can handle which addresses the incompleteness for when `Tx` is a type variable. At the same time we can encode negated types using `TypeSwitch(P, nothing, _)` allowing us to precisely represent negations for any type predicate `P`, addressing the second cause of incompleteness

# Code Decisions

This diff introduce the `TypeSwitch` constraint type, but doesn't specify any additional subtyping logic. This will be introduced in subsequent diffs. This is mostly boilerplate code that is needed whenever we add a new constraint type and thought it would be easier to review if these parts were separated from the subtyping rules.

Of significance for this diff, `TypeSwitch` is very similar to a union type `T1 | T2` where we know `T1` and `T2` are disjoint. Because of this, the variance of `T1` and `T2` should be covariant motivating the change in `Typing_env`

Also I've introduced the type `Typing_defs_core.type_predicate`. Right now it only encodes `IsBool`, but more predicates will be supported once this initial predicate is made to work.

Differential Revision: D54664466

fbshipit-source-id: 63e7b8271b18e1055de16830cfc2983607b68bfd