unify ir::SrcLoc + hhbc::SrcLoc
Summary:
Many ir and hhbc types contain a SrcLoc, so we need to unify SrcLoc
before much else can be unified. ir::SrcLoc contains a filename, hhbc::SrcLoc doesn't.

We require hhbc<->IR conversions to be semantics preserving, but we can't preserve
the filename field without adding it to hhbc::SrcLoc. Similarly, HHVM structures would
need to have a filename field, or else what's the point?

If/when hackc begins inlining or cross-file const propagation, it might be necessary
to put filenames in srcloc so they are unambiguous. But we'll have to do that all the
way through HHVM for it to be useful.

infer, pivv, and privacy_profilers process one IR unit at a time and know the filename.

Conclusion: Remove ir::SrcLoc::filename, for now, then use hhbc::SrcLoc everywhere.

Reviewed By: aorenste

Differential Revision: D54221175

fbshipit-source-id: cb2e4d0862a6dd8ade59b3bfb899a222cdd272b0