Introduce `FOLLY_HAS_JEMALLOC_DEFS`
Summary:
jemalloc headers define constants that can be used from user code. One
example of such constant is `MALLCTL_ARENAS_ALL`.

Currently, if we want to use such constants we need to include jemalloc header,
but it is extremely tricky to determine if we actually can include this header,
when we are dealing with some low level library. Basically, to make it work
reliable we need to replicate the same logic folly/portability/Malloc.h already
has.

Let's explicitly define `FOLLY_HAS_JEMALLOC_DEFS`, when we can and
actually do include jemalloc headers.

Reviewed By: ot, interwq

Differential Revision: D54754592

fbshipit-source-id: cc0f1cc68516b0feb342f6258d5c22bd7226e28f