Collapse use of ::types::
Summary:
This diff is part of the codemod for https://fb.workplace.com/groups/rust.language/posts/25780668961555042.

Thrift-generated data structures, clients, and service traits all used to be exported in a single crate, with the data structures being under a `types` module. Now that data structures have a crate of their own, the `types` module is vestigial: `foo::types` refers to the same contents as `foo`.

Reviewed By: zertosh

Differential Revision: D54871740

fbshipit-source-id: f7955cf0e08291c6e6758bbd90dcd6b3b8ada7a0