Collapse use of ::types:: (T-V)
Summary:
This diff is part of the codemod for https://fb.workplace.com/groups/rust.language/posts/25780668961555042.

Thrift-generated data structures, clients, and service traits all used to be exported in a single crate, with the data structures being under a `types` module. Now that data structures have a crate of their own, the `types` module is vestigial: `foo::types` refers to the same contents as `foo`.

Generated using the code in P1195926494 and the procedure from D54838520.

Reviewed By: zertosh

Differential Revision: D54871744

fbshipit-source-id: a70331f5df5dad7e81e8d25ceb25af5edd00d388