new method in ext_facts, getFilesAndAttrValsWithAttribute
Summary:
There is a workflow used by ent gencode we would like to support that involves getting all the files with a particular file attr, and all the vals for that file attr for each file.

Right now it is done with one HH\facts query to get the files, then N HH\facts queries to get the attribute vals for each one. This is slow.  The new API allows us to do this in one step and should improve the performance of gencode.

There's a weird behavior here --  if you supply a file attr arg as an int, it comes back out as a string. This is not new to my code, it agrees with the existing getFileAttrArgs method.  But it's strange and so I am mentioning it (and i also mention it in a comment.)

Reviewed By: edwinsmith

Differential Revision: D54715253

fbshipit-source-id: a94970b5541e0315c46b2b873296786d9201bbc2